refactor: изменение архитектуры
