--- app/frontend/admin/pages/statistics.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Статистика</title>
-    <link rel="stylesheet" href="../css/style.css">
-</head>
-<body>
-<div class="header">
-    <nav><a href="statistics.html">Статистика</a>  </nav>
-    <nav><a href="editor.html"> Каталог</a></nav>
-    <nav><a href="backup.html"> Бэкап </a></nav>
-</div>
-тут будет статистика
-</body>
-</html>
