--- app/frontend/home/css/style.css
@@ -1,50 +1,0 @@
-.film-grid {
-    display: grid;
-    grid-template-columns: repeat(5, 1fr);
-    gap: 20px;
-    padding: 20px;
-}
-
-.film-card {
-    cursor: pointer;
-    transition: transform 0.3s;
-    background: #222;
-    border-radius: 8px;
-    overflow: hidden;
-    color: white;
-}
-
-.film-card:hover {
-    transform: scale(1.05);
-}
-
-.film-card img {
-    width: 100%;
-    height: 300px;
-    object-fit: cover;
-}
-
-.film-card h3 {
-    margin: 10px;
-    font-size: 16px;
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-}
-
-.film-card p {
-    margin: 0 10px 10px;
-    color: #aaa;
-}
-
-.menu__link {
-    color: white;
-    text-decoration: none;
-    padding: 5px 10px;
-}
-
-.menu__link_active {
-    color: red;
-    font-weight: bold;
-    border-bottom: 2px solid red;
-}
