+++ app/static/admin/js/editor.js
