+++ app/static/assets/images/search-icon.png
