+++ app/templates/editor.html
@@ -1,9 +1,9 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="ru">
 <head>
     <meta charset="UTF-8">
     <title>Редактор каталогов</title>
-    <link rel="stylesheet" href="../css/style.css">
+    <link rel="stylesheet" href="{{ url_for('static', filename='admin/css/style.css') }}">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css">
 </head>
 <body>
@@ -19,7 +19,7 @@
 <div id="films-container" class="films-container">
 
 </div>
-<script src="../js/editor.js"></script>
+<script src="{{ url_for('static', filename='admin/js/editor.js') }}"></script>
 
 </body>
 </html>
