+++ app/templates/home.html
@@ -2,26 +2,26 @@
 <html lang="ru">
 <head>
     <meta charset="UTF-8">
-    <title>Главная страница</title>
-    <link rel="stylesheet" href="../css/style.css">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Главная страница</title>
+    <link rel="stylesheet" href="{{ url_for('static', filename='home/css/style.css') }}">
 </head>
 <body class="root">
     <main class="content">
         <div class="search">
             <input id="searchInput" class="search__input" name="search" type="text" placeholder="Поиск">
             <button id="searchButton" class="search__button">
-                <img class="search__icon" src="../../assets/images/search-icon.png">
+                <img class="search__icon" src="{{ url_for('static', filename='assets/images/search-icon.png') }}">
             </button>
         </div>
         <div class="menu-container">
             <button class="sorting"></button>
             <nav class="menu">
                 <ul>
-                    <li><a href="home.html" class="menu__link menu__link_active">Популярно сейчас</a></li>
-                    <li><a href="home.html" class="menu__link">Фильмы</a></li>
-                    <li><a href="home.html" class="menu__link">Сериалы</a></li>
-                    <li><a href="recomendations.html" class="menu__link">Рекомендации</a></li>
+                    <li><a href="#" class="menu__link menu__link_active">Популярно сейчас</a></li>
+                    <li><a href="#" class="menu__link">Фильмы</a></li>
+                    <li><a href="#" class="menu__link">Сериалы</a></li>
+                    <li><a href="#" class="menu__link">Рекомендации</a></li>
                 </ul>
             </nav>
             <button class="filter"></button>
@@ -45,6 +45,6 @@
         </div>
     </template>
 
-    <script src="/home/js/main.js"></script>
+    <script src="{{ url_for('static', filename='home/js/home.js') }}"></script>
 </body>
 </html>
