+++ app/templates/login.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>Вход в админ-панель</title>
-    <link rel="stylesheet" href="../css/style.css">
+    <link rel="stylesheet" href="{{ url_for('static', filename='admin/css/style.css') }}">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
 </head>
