+++ app/templates/recomendations.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="ru">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Рекомендации</title>
+    <link rel="stylesheet" href="{{ url_for('static', filename='home/css/style.css') }}">
+</head>
+<body>
+    <p>Рекомендации</p>
+</body>
+</html>
