+++ app/templates/statistics.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="ru">
+<head>
+    <meta charset="UTF-8">
+    <title>Статистика</title>
+    <link rel="stylesheet" href="{{ url_for('static', filename='admin/css/style.css') }}">
+</head>
+<body>
+<div class="header">
+    <nav><a href="#">Статистика</a>  </nav>
+    <nav><a href="{{ url_for(editor.html) }}"> Каталог</a></nav>
+    <nav><a href="{{ url_for(backup.html) }}"> Бэкап </a></nav>
+</div>
+тут будет статистика
+</body>
+</html>
